requirements.txt
undetected-chromedriver
selenium
requests
beautifulsoup4
setuptools
python-dotenv